+++ pmd-plsql/src/main/java/net/sourceforge/pmd/lang/plsql/rule/design/AbstractNcssCountRule.java
@@ -2,7 +2,7 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.lang.plsql.rule.codesize;
+package net.sourceforge.pmd.lang.plsql.rule.design;
 
 import java.util.logging.Level;
 import java.util.logging.Logger;
