+++ pmd-plsql/src/main/java/net/sourceforge/pmd/lang/plsql/rule/design/ExcessiveObjectLengthRule.java
@@ -2,10 +2,9 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.lang.plsql.rule.codesize;
+package net.sourceforge.pmd.lang.plsql.rule.design;
 
 import net.sourceforge.pmd.lang.plsql.ast.OracleObject;
-import net.sourceforge.pmd.lang.plsql.rule.design.ExcessiveLengthRule;
 
 /**
  * This rule detects when an Oracle object exceeds a certain threshold. i.e. if
