+++ pmd-plsql/src/test/resources/net/sourceforge/pmd/lang/plsql/rule/design/xml/NPathComplexity.xml
